Fix : Float타입 Double로 변경

+++ watch/app/src/main/java/com/runapp/model/RunDetail.java
@@ -2,13 +2,13 @@ package com.runapp.model;
 
 public class RunDetail {
     private Long second;   // 해당 기록이 기록된 시간
-    private float speed; // 속도(m/s)
+    private double speed; // 속도(m/s)
     private String pace; // 페이스(m/k)
-    private float distance; // 거리(m)
-    private float heartRate; // 심박수
+    private double distance; // 거리(m)
+    private double heartRate; // 심박수
     private String state; // 상태
-    private float latitude; // 위도
-    private float longitude; // 경도
+    private double latitude; // 위도
+    private double longitude; // 경도
 
     public Long getSecond() {
         return second;
@@ -18,11 +18,11 @@ public class RunDetail {
         this.second = second;
     }
 
-    public float getSpeed() {
+    public double getSpeed() {
         return speed;
     }
 
-    public void setSpeed(float speed) {
+    public void setSpeed(double speed) {
         this.speed = speed;
     }
 
@@ -34,19 +34,19 @@ public class RunDetail {
         this.pace = pace;
     }
 
-    public float getDistance() {
+    public double getDistance() {
         return distance;
     }
 
-    public void setDistance(float distance) {
+    public void setDistance(double distance) {
         this.distance = distance;
     }
 
-    public float getHeartRate() {
+    public double getHeartRate() {
         return heartRate;
     }
 
-    public void setHeartRate(float heartRate) {
+    public void setHeartRate(double heartRate) {
         this.heartRate = heartRate;
     }
 
@@ -58,19 +58,19 @@ public class RunDetail {
         this.state = state;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
