+++ watch/app/src/main/java/com/runapp/model/RunningViewModel.java
@@ -7,35 +7,35 @@ import androidx.lifecycle.ViewModel;
 
 public class RunningViewModel extends ViewModel {
 
-    private final MutableLiveData<Float> heartRate = new MutableLiveData<>();
+    private final MutableLiveData<Double> heartRate = new MutableLiveData<>();
     private final MutableLiveData<String> elapsedTime = new MutableLiveData<>();
     private final MutableLiveData<String> msPace = new MutableLiveData<>();
-    private final MutableLiveData<Float> msSpeed = new MutableLiveData<>();
-    private final MutableLiveData<Float> distance = new MutableLiveData<>();
-    private final MutableLiveData<Float> stepCounter = new MutableLiveData<>();
-    private final MutableLiveData<Float> latitude = new MutableLiveData<>();
-    private final MutableLiveData<Float> longitude = new MutableLiveData<>();
+    private final MutableLiveData<Double> msSpeed = new MutableLiveData<>();
+    private final MutableLiveData<Double> distance = new MutableLiveData<>();
+    private final MutableLiveData<Double> stepCounter = new MutableLiveData<>();
+    private final MutableLiveData<Double> latitude = new MutableLiveData<>();
+    private final MutableLiveData<Double> longitude = new MutableLiveData<>();
     private MutableLiveData<RunningData> runningData = new MutableLiveData<>();
 
     public MutableLiveData<RunningData> getRunningData() {
         return runningData;
     }
-    public void setHeartRate(float rate) {
+    public void setHeartRate(double rate) {
         heartRate.setValue(rate);
     }
 
-    public LiveData<Float> getHeartRate() {
+    public LiveData<Double> getHeartRate() {
         return heartRate;
     }
 
-    public void setDistance(float value) {
+    public void setDistance(double value) {
         distance.setValue(value);
     }
     public void setMsPace(String value) {
         msPace.setValue(value);
     }
 
-    public LiveData<Float> getDistance() {
+    public LiveData<Double> getDistance() {
         return distance;
     }
 
@@ -51,7 +51,7 @@ public class RunningViewModel extends ViewModel {
         return msPace;
     }
 
-    public MutableLiveData<Float> getMsSpeed() {
+    public MutableLiveData<Double> getMsSpeed() {
         return msSpeed;
     }
 
@@ -59,29 +59,29 @@ public class RunningViewModel extends ViewModel {
         this.runningData = runningData;
     }
 
-    public void setStepCounter(float value) {
+    public void setStepCounter(double value) {
         this.stepCounter.setValue(value);
     }
 
-    public void setMsSpeed(float value) {
+    public void setMsSpeed(double value) {
         this.msSpeed.setValue(value);
     }
 
-    public LiveData<Float> getStepCounter() {
+    public LiveData<Double> getStepCounter() {
         return stepCounter;
     }
 
-    public MutableLiveData<Float> getLatitude() {
+    public MutableLiveData<Double> getLatitude() {
         return latitude;
     }
 
-    public MutableLiveData<Float> getLongitude() {
+    public MutableLiveData<Double> getLongitude() {
         return longitude;
     }
-    public void setLatitude(float latitude){
+    public void setLatitude(double latitude){
         this.latitude.setValue(latitude);
     }
-    public void setLongitude(float longitude){
+    public void setLongitude(double longitude){
         this.longitude.setValue(longitude);
     }
 }
